+++ src/autostore/__init__.py
@@ -1,6 +1,6 @@
 """autostore."""
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 
 from . import fetch, qc, write
 from .calcn import Calculation