+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 ### Added
 - Geometry hash field populated by event listener
+- Calculation object with hash registry
+- Calculation hash table populated by event listener
+- QCIO ProgramInput generation from Calculation object
 
 ## [0.0.2] - 2026-01-28
 ### Added