+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.0.4] - 2026-04-01
+
 ## [0.0.3] - 2026-01-29
 ### Added
 - Geometry hash field populated by event listener